digiplot/tests/test_plot_area.py
import unittest
# import unittest2 as unittest # for versions of python < 2.7

"""
        Method                  Checks that
self.assertEqual(a, b)           a == b   
self.assertNotEqual(a, b)        a != b   
self.assertTrue(x)               bool(x) is True  
self.assertFalse(x)              bool(x) is False     
self.assertIs(a, b)              a is b
self.assertIsNot(a, b)           a is not b
self.assertIsNone(x)             x is None 
self.assertIsNotNone(x)          x is not None 
self.assertIn(a, b)              a in b
self.assertNotIn(a, b)           a not in b
self.assertIsInstance(a, b)      isinstance(a, b)  
self.assertNotIsInstance(a, b)   not isinstance(a, b)  

See:
      https://docs.python.org/2/library/unittest.html
         or
      https://docs.python.org/dev/library/unittest.html
for more assert options
"""

import sys, os

here = os.path.abspath(os.path.dirname(__file__)) # Needed for py.test
up_one = os.path.split( here )[0]  # Needed to find digiplot development version
if here not in sys.path[:2]:
    sys.path.insert(0, here)
if up_one not in sys.path[:2]:
    sys.path.insert(0, up_one)

from digiplot.plot_area import PlotArea

class MyTest(unittest.TestCase):

    def setUp(self):
        unittest.TestCase.setUp(self)
        self.myclass = PlotArea(w_canv=1000, h_canv=1000)
        self.myclass.open_img_file( os.path.join(here,'Cu_cte.png') )

    def tearDown(self):
        unittest.TestCase.tearDown(self)
        del( self.myclass )

    def test_should_always_pass_cleanly(self):
        """Should always pass cleanly."""
        pass

    def test_image_size(self):
        C = self.myclass
        self.assertEqual( (792, 612), (C.w_img,C.h_img) )

    def test_myclass_existence(self):
        C = self.myclass
        self.assertTrue(C)
        self.assertEqual(C.fi_origin, 0.0)
        self.assertEqual(C.fj_origin, 1.0)

    def test_set_origin_xy(self):
        C = self.myclass
        C.set_origin_xy( 10., 1000. )
        self.assertEqual( (C.x_origin,C.y_origin), (10., 1000.))


if __name__ == '__main__':
    # Can test just this file from command prompt
    #  or it can be part of test discovery from nose, unittest, pytest, etc.
    unittest.main()

